Made accept and cancel features for lot, schedule for starting or closing lot

+++ src/main/java/com/epam/au/dao/impl/UserDataBaseDAO.java
@@ -209,7 +209,7 @@ public class UserDataBaseDAO implements DataBaseDAO {
         try {
             connection = connectionPool.takeConnection();
             statement = connection.prepareStatement(queryBundle.getQuery("select.any.by_id.roles"));
-            statement.setDouble(1, id);
+            statement.setLong(1, id);
 
             rs = statement.executeQuery();
